[FIX] website_sale: set confirmation mail template setting to readonly
In website_sale setting, there is a field 'Confirmation Email' that is not used,
and which purpose is to have a conveninent link to edit the mail templates.
Since it is an editable setting, this confuses the users.
It was in readonly before commit:
https://github.com/odoo/odoo/commit/bccbd00dc3e6448918aceb28bb0931e3ee336db9#diff-3f25fe0e27881e9c6655bbb6974c84abR59
which removed the readonly without any apparent reason.

We put it back into readonly mode.

opw 1860167