[FIX] stock: be able to update stock module (demo data)
Due to commit d9913cf0316da10253d1d8ecfe077239d33341cd when a user
with demo data try to update stock it will return the location
constraint because the demo data try to update the supplier location
with the usage supplier. However there is no check if the usage is
really modify or just in values.

This commit add a check in order to only select locations with a new
usage or that effectively change the 'scrap location' switch.