[FIX] stock, delivery: packaging_id displayed in delivery
The field 'packaging_id' of quant package was displayed in tree view,
even if it was impossible to fullfill it because it is added in the form
in module delivery. So the field in the tree view is moved in the module
delivery to be shown whan you can set it.