[FIX] sale_payment: Pay Now was always hidden in SO panel-heading
Since refactoring of account, payment and sale with commit dcb6dca3 the
`Pay Now` button in the panel-heading of the SO in the portal was always hidden
except if there was a pending transaction.

The button should be visible when there is no transaction. In that case
`payment_tx_id` is an empty recordset, we can safely access its `state` field.

Closes #25529