[FIX] mail:tests: more precise jQuery selector on modal body content
Tests from [1] were not passing, due to changes from [2].

This is due to the fact that `$.('.modal-body')` picks 2 elements, instead of 1.

Partial HTML before [2]:

```
    <main class="modal-body">
        <div>
            <p>
                ...
            </p>
        </div>
    </main>
```

Partial HTML after [1]:

```
    <main class="modal-body">
        <main role="alert" class="modal-body">
            <p>
                ...
            </p>
        </main>
    </main>
```

In order to not pick both DOM elements, we refine the jQuery selector to pick
the `<p>`, so that there's only one element and it has the same text content.

[1] [FIX] mail: warn user on record reload before post message
    https://github.com/odoo/odoo/commit/655d4d8a9920db4af8bcd7b7cbdcc395a6cc51d5
[2] [IMP] Improve compatibility with screen readers (accessibility) (#24574)
    https://github.com/odoo/odoo/commit/b80ad4f59227902a077236fcb9e1b3489954e527